+++ src/com/interrupt/dungeoneer/screens/MainMenuScreen.java
@@ -336,9 +336,9 @@ public class MainMenuScreen
         } else if (Gdx.input.isKeyJustPressed(Input.Keys.F5)) {
             GameApplication.SetScreen(new MainMenuScreen());
         } else if (Gdx.input.isKeyJustPressed(Input.Keys.FORWARD_DEL)) {
-                if(selectedSave != null && saveGames[selectedSave] != null && !isSelected) {
-                    GameApplication.SetScreen(new ConfirmScreen(selectedSave));
-                }
+            if(selectedSave != null && saveGames[selectedSave] != null && !isSelected) {
+                GameApplication.SetScreen(new ConfirmScreen(selectedSave));
+            }
         }
 
         this.ui.act(delta);