+++ src/net/cotd/delverunlimited/Config.java
@@ -10,7 +10,7 @@ public class Config {
     public static boolean skipIntro = false;
     public static int forceSave = 3;
 
-    public static String OfflineVer = "1.5.0";
+    public static String OfflineVer = "1.6.0";
     public static boolean isRelease = true; // is this a RTM build or just a test build?
 
     public static int height = 0;