+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Fixed erasing a save which already is empty
+
 ## [1.5.0] - 2017-08-31
 ### Added
 - Double tapping a save will now load it in the main menu